requirements.txt
# ─────────────────────────────────────────────
#  Voice AI Agent — Python Dependencies
# ─────────────────────────────────────────────

# Core framework
pipecat-ai[deepgram,openai,elevenlabs,silero,livekit]==0.0.47

# Web server (webhook handler)
fastapi==0.111.0
uvicorn[standard]==0.30.1
python-multipart==0.0.9

# Twilio SDK
twilio==9.2.3

# LiveKit SDK
livekit==0.11.1
livekit-api==0.6.0

# Azure OpenAI
openai==1.35.3

# Knowledge base
chromadb==0.5.3
sentence-transformers==3.0.1

# Metrics / observability
prometheus-client==0.20.0
prometheus-fastapi-instrumentator==7.0.0

# Resilience
tenacity==8.4.1          # retry + backoff
circuitbreaker==2.0.0    # circuit breaker

# Utilities
python-dotenv==1.0.1
pydantic==2.7.4
pydantic-settings==2.3.4
structlog==24.2.0         # structured logging
httpx==0.27.0
asyncio-pool==0.6.0
numpy==1.26.4
aiofiles==23.2.1
